Apply ContainerLifecycleHooks to KafkaContainer
